requirements.txt
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
lxml==5.1.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
numpy==1.26.4
pandas==2.2.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
requests==2.31.0
simple-websocket==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
WTForms==3.1.1
